// src/web/src/api/generated/models/SaveSessionCheckinRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type SaveSessionCheckinRequest = {
    sessionRpe: number | string | null;
    pain: number | string | null;
    painLocation: string | null;
    gaitChanged: boolean | null;
    fatigue: number | string | null;
    sleepQuality: number | string | null;
    perceivedRecovery: number | string | null;
    hasIllnessOrSymptom: boolean | null;
    symptomNote: string | null;
    recoveryResponse: string | null;
    note: string | null;
};