// src/web/src/api/generated/models/SessionActivityResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type SessionActivityResponse = {
    id: string;
    garminActivityId: number | string | null;
    title: string | null;
    activityType: string;
    activityCategory: string | null;
    modality: string | null;
    startedAtLocal: string;
    distanceM: number | string | null;
    durationSeconds: number | string | null;
};